Refactor BitBarDevices
