+++ bin/maze-runner
@@ -8,15 +8,15 @@ require_relative '../lib/utils/deep_merge'
 require_relative '../lib/maze'
 
 require_relative '../lib/maze/appium_server'
-require_relative '../lib/maze/bitbar_devices'
 require_relative '../lib/maze/bugsnag_config'
 require_relative '../lib/maze/client/bb_client_utils'
 require_relative '../lib/maze/client/bs_client_utils'
 require_relative '../lib/maze/client/appium/base_client'
 require_relative '../lib/maze/client/appium/bb_client'
+require_relative '../lib/maze/client/appium/bb_devices'
 require_relative '../lib/maze/client/appium/bs_client'
-require_relative '../lib/maze/client/appium/local_client'
 require_relative '../lib/maze/client/appium/bs_devices'
+require_relative '../lib/maze/client/appium/local_client'
 require_relative '../lib/maze/compare'
 require_relative '../lib/maze/docker'
 require_relative '../lib/maze/document_server'
